Fixed scripts, Added verbose output
Fixes #29

+++ bin/backup.sh
@@ -19,7 +19,12 @@ case "$1" in
     ## MySQL
     ###################################
     "mysql")
-        rm -f -- "${BACKUP_DIR}/${BACKUP_MYSQL_FILE}"
+        if [ -f "${BACKUP_DIR}/${BACKUP_MYSQL_FILE}" ]; then
+            logMsg "Removing old backup file..."
+            rm -f -- "${BACKUP_DIR}/${BACKUP_MYSQL_FILE}"
+        fi
+
+        logMsg "Starting MySQL backup..."
         mysqldump --opt --all-databases | bzip2 > "${BACKUP_DIR}/${BACKUP_MYSQL_FILE}"
         ;;
 
@@ -27,7 +32,12 @@ case "$1" in
     ## Solr
     ###################################
     "solr")
-        rm -f -- "${BACKUP_DIR}/${BACKUP_SOLR_FILE}"
-        tar jcf "${BACKUP_DIR}/${BACKUP_SOLR_FILE}" /data/solr/
+        if [ -f "${BACKUP_DIR}/${BACKUP_SOLR_FILE}" ]; then
+            logMsg "Removing old backup file..."
+            rm -f -- "${BACKUP_DIR}/${BACKUP_SOLR_FILE}"
+        fi
+
+        logMsg "Starting Solr backup..."
+        tar jcPf "${BACKUP_DIR}/${BACKUP_SOLR_FILE}" /data/solr/
         ;;
 esac
